LogMatcher returns LogLine
